2024/3/18/0:09

+++ models/sequential/dcrec_seq.py
@@ -151,7 +151,7 @@ def graph_augment(g: dgl.DGLGraph, user_ids, user_edges):
         node_indicies_a).to(g.device)
     node_indicies_b = torch.from_numpy(
         node_indicies_b).to(g.device)
-    edge_ids = g.edge_ids(node_indicies_a, node_indicies_b)
+    edge_ids = g.edge_ids(node_indicies_a.to(torch.int64), node_indicies_b.to(torch.int64))
 
     aug_g: dgl.DGLGraph = deepcopy(g)
     # The features for the removed edges will be removed accordingly.
